[SPARK-43884] Param markers in DDL
### What changes were proposed in this pull request?

In this change we allow parameter markers (:parm) to be used in various DDL statements, specifically as input to the IDENTIFIER-clause to allow templating of statements such as CREATE TABLE IDENTIFIER(:mytab).
We block parameter markers from view bodies to prevent them from appearing in the persisted view definition.

### Why are the changes needed?

This change is needed to allow exploitation of the IDENTIFIER clause in DDL statements

### Does this PR introduce _any_ user-facing change?

Yes

### How was this patch tested?

Added new tests to the parameter suite.

Closes #41395 from srielau/SPARK-43884-param-markers-in-ddl.

Lead-authored-by: srielau <serge@rielau.com>
Co-authored-by: Gengliang Wang <gengliang@apache.org>
Co-authored-by: Serge Rielau <srielau@users.noreply.github.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>