[SPARK-43798][PYTHON][FOLLOW-UP] Sets the batch size as 1 to send individual group
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/41316 that sets the correct batch size as 1 when it streams from Python to JVM.

In addition, this PR contains slight cleanup.

### Why are the changes needed?

Currently we're sending 100 batches instead of 100 rows within a batch when we output the results, meaning that we will send 100 iterators from 100 UDTF invocations with different arguments. This is because one row in `BatchedSerializer` is actually one group from one UDTF invocation within a tuple. This PR sets the batch size as 1 for now.

Ideally we should implement its custom logic to batch them per row-size. However, this requires to define an additional protocol (e.g., to signal which is the end of one UDTF invocation/batch). Therefore, this PR fixes it with a minimized change for now.

### Does this PR introduce _any_ user-facing change?

No, UDTF is not released to the end users yet.

### How was this patch tested?

Manually tested.

Closes #41563 from HyukjinKwon/SPARK-43798-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>