[SPARK-44012][SS] KafkaDataConsumer to print some read status
### What changes were proposed in this pull request?
In the end of each KafkaDataConsumer, it logs some stats. Here is an sample log line:

23/06/08 23:48:14 INFO KafkaDataConsumer: From Kafka topicPartition=topic-121-2 groupId=spark-kafka-source-623fa0a8-04a5-4f34-a9ad-adbf31494e85-711383366-executor read 1 records, taking 504554479 nanos, during time span of 504620999 nanos

### Why are the changes needed?
For each task, Kafka source should report fraction of time spent in KafkaConsumer to fetch records. It should also report overall read bandwidth (bytes or records read / time spent fetching).

This will be useful in verifying if fetching is the bottleneck.

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
1. Run unit tests and validate log line is correct
2. Run some benchmarks and see it doesn't show up much in CPU profiling.

Closes #41525 from siying/kafka_logging2.

Authored-by: Siying Dong <siying.dong@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>