[SPARK-42941][SS][CONNECT] 1/2] StreamingQueryListener - Event Serde in JSON format
### What changes were proposed in this pull request?

Following the discussion of `foreachBatch` implementation, we decide to implement connect StreamingQueryListener in a way that the server runs the listener code, rather than the client.

Following this POC: https://github.com/apache/spark/pull/41096, this is going to be done in a way such that
1. Client sends serialized python code to server
2. Server initializes a Scala `StreamingQueryListener`, which initialize the python progress and run the python code. (Details of this step still depends on `foreachBatch` implementation.
3. When a new StreamingQuery Event comes in, the jvm serialize it to JSON and send it to the python progress to process.

This PR focus on step 3, the serialization and deserialization of the events.

Also finishes a TODO to check exception in QueryTerminatedEvent

### Why are the changes needed?

For implementing Connect StreamingQueryListener

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

New unit tests

Closes #41540 from WweiL/SPARK-42941-listener-python-new-1.

Authored-by: Wei Liu <wei.liu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>