[SPARK-44136][SS] Fixed an issue that StateManager may get materialized in executor instead of driver in FlatMapGroupsWithStateExec
### What changes were proposed in this pull request?

Fixed an issue that StateManager may get materialized in executor instead of driver in FlatMapGroupsWithStateExec. The ticket that brought this issue: https://issues.apache.org/jira/browse/SPARK-40411

The basic idea is to maintain the `stateManager` as `lazy val` but initialize it earlier in the `doExecute()` to force a lazy init at driver.

### Why are the changes needed?

Because without this change, the StateManager in FlatMapGroupsWithStateExec may get materialized in executor instead of driver which would cause unexpected behavior.

### Does this PR introduce _any_ user-facing change?

Yes

### How was this patch tested?

It's hard to write a unit test for this since it involves in both driver and executor which is hard to simulate through a unit test.

Closes #41693 from bogao007/SPARK-44136.

Authored-by: bogao007 <bo.gao@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>