[SPARK-43136][CONNECT][FOLLOWUP] Adding tests for KeyAs
### What changes were proposed in this pull request?
The current impl of KeyAs for the Scala client is a purely client side encoder operation. Thus we could end up with duplicates in keys.

Added the tests both for the Scala Client and for Spark dataset API. It showed the behavior is the same for server and client at this moment.

### Why are the changes needed?
More tests to verify the client and server behavior.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Tests

Closes #40980 from zhenlineo/keys.

Authored-by: Zhen Li <zhenlineo@users.noreply.github.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>