[SPARK-43922][SQL] Add named parameter support in parser for function calls
### What changes were proposed in this pull request?
We plan on adding two new tokens called ```namedArgumentExpression``` and ```functionArgument``` which would enable this feature. When parsing this logic, we also make changes to ASTBuilder such that it can detect if the argument passed is a named argument or a positional one.

Here is the link for the design document:
https://docs.google.com/document/d/1uOTX0MICxqu8fNanIsiyB8FV68CceGGpa8BJLP2u9o4/edit

### Why are the changes needed?
This is part of a larger project to implement named parameter support for user defined functions, built-in functions, and table valued functions.

### Does this PR introduce _any_ user-facing change?
Yes, the user would be able to call functions with argument lists that contain named arguments.

### How was this patch tested?
We add tests in the PlanParserSuite that will verify that the plan parsed is as intended.

Closes #41796 from learningchess2003/43922-new.

Authored-by: Richard Yu <richard.yu@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>