[SPARK-44291][SPARK-43416][CONNECT] Fix incorrect schema for range query
### What changes were proposed in this pull request?

Configure the spark connect client to use the schema
returned from the service for all requests, instead of
assuming the schema for data sets.

### Why are the changes needed?

The `range()` query returns a result with one column named "id".
However, when the underlying schema is queried, it returns the schema
to contain one column named "value".

This is a bug and is different from the behaviour in other places,
including pyspark.

Incidentally, this also fixes SPARK-43416.

### Does this PR introduce _any_ user-facing change?

Yes. It fixes the bug for incorrect column names returned for a
range query.

### How was this patch tested?

Tests attached.

Closes #41846 from nija-at/fix-range-schema.

Authored-by: Niranjan Jayakar <nija@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>