[SPARK-44282][CONNECT] Prepare DataType parsing for use in Spark Connect Scala Client
### What changes were proposed in this pull request?
This PR prepares moving DataType parsing to sql/api. Basically it puts all DataType parsing functionality in a super class of regular parsing. We cannot move the parser just yet because that need to happen at the same time as DataType.

### Why are the changes needed?
We want the Spark Connect Scala Client to use a restricted class path. DataType will be one of the shared classes, to move DataType we need to move DataType parsing.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Modified existing tests.

Closes #41836 from hvanhovell/SPARK-44282.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>