[SPARK-44275][CONNECT] Add configurable retry mechanism to Scala Spark Connect
### What changes were proposed in this pull request?
This PR introduces a configurable retry mechanism for the Scala `SparkConnectClient`.
Parameters for the exponential backoff and a filter for exceptions to retry are passed to the client via the existing (extended) `Configuration` class. By default, only RuntimeStatusException: Unavailable triggers a retry - this mirrors the default retry behavior in the Python client.

One might want to move the retry logic into the GRPC stub that will potentially be introduced [here](https://github.com/apache/spark/pull/41743).

### Why are the changes needed?
There are a few existing exceptions that one might want to handle with a retry.
For example, this would allow one to not exit with an exception when executing a command while the cluster is still starting.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Tests included.

Closes #41829 from dillitz/spark-44275-retries.

Authored-by: Robert Dillitz <robert.dillitz@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>