[SPARK-44312][CONNECT][PYTHON] Allow to set a user agent with an environment variable
### What changes were proposed in this pull request?
Use the `SPARK_CONNECT_USER_AGENT` environment variable as a fallback for the prepended user agent string if one did not set the user agent in `ChannelBuilder`.

### Why are the changes needed?
Currently one has to specify a custom user agent string in `ChannelBuilder`. It would be useful to be able to set this string with an environment variable.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Manual testing + existing tests.

Closes #41866 from dillitz/SPARK-44312-user-agent-environment.

Authored-by: Robert Dillitz <robert.dillitz@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>