[SPARK-44352][CONNECT] Put back sameType and friends in DataType
### What changes were proposed in this pull request?
This PR moves back sameType and friends into DataType. Now that we have a conf system for SQL/API we can host them in there.

### Why are the changes needed?
We want to retain compatibility with the existing DataType APIs. Also we need to eliminate the use of catalyst specific code in the DataType hierarchy.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing tests.

Closes #41921 from hvanhovell/SPARK-44352.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>