[SPARK-44399][PYHTON][CONNECT] Import SparkSession in Python UDF only when useArrow is None
### What changes were proposed in this pull request?
Import SparkSession in Python UDF only when useArrow is None.

When useArrow is None, SparkSession is used to check the value of the Spark Conf "spark.sql.execution.pythonUDF.arrow.enabled".

Delaying the import only when useArrow is None enables users to define a Python UDF without a SparkSession when (useArrow is not None).

### Why are the changes needed?
Better usability.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Manual tests.

Closes #41970 from xinrong-meng/fix_import.

Authored-by: Xinrong Meng <xinrong@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>