[SPARK-44390][CORE][SQL] Rename `SparkSerDerseUtils` to `SparkSerDeUtils`
### What changes were proposed in this pull request?
The change to this pr is to rename `SparkSerDerseUtils` to `SparkSerDeUtils`.

### Why are the changes needed?
`SerDe` is a more commonly used naming convention in Spark, such as `o.a.s.api.python.SerDeUtil`, `o.a.s.ml.python.MLSerDe` and `o.a.s.sql.internal.HiveSerDe`, etc., so we should follow this habit.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass Git Hub Actions

Closes #41960 from LuciferYang/SPARK-44390.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>