[SPARK-42454][SQL] SPJ: encapsulate all SPJ related parameters in BatchScanExec
### What changes were proposed in this pull request?
Pull out the SPJ-related attribute of BatchScanExec into a case class

### Why are the changes needed?
We plan to have further evolution of SPJ parameters to support more SPJ features.  So we want to stabilize the definition of BatchScanExec to not have to touch the many places in the code that it is pattern-matched/unapplied, etc..

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing unit test to verify no behavior change.

Closes #41990 from szehon-ho/spj_refactor.

Authored-by: Szehon Ho <szehon.apache@gmail.com>
Signed-off-by: Chao Sun <sunchao@apple.com>