[SPARK-44348][CONNECT][FOLLOW-UP] Avoid double slashes in the URI
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/41942 that does `substring(1)` to remove the leading slash so it makes it relative parts from URI. Otherwise, it can end up with having double slashes in the middle.

### Why are the changes needed?

To avoid having unnecessary double slashes  ... and save one byte :-)

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manually tested. It's really trivial.

Closes #42051 from HyukjinKwon/minor-change.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>