[SPARK-44481][CONNECT][PYTHON] Make pyspark.sql.is_remote an API
### What changes were proposed in this pull request?

This PR proposes to expose and document `pyspark.sql.is_remote` as an API.

### Why are the changes needed?

For the end users to be able to do if-else, e.g., for dispatching the code path to the legacy mode or connect mode.

### Does this PR introduce _any_ user-facing change?

Yes, it exposes a method as an API.

### How was this patch tested?

Manually built and checked the documentation.

Closes #42072 from HyukjinKwon/SPARK-44481.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>