[SPARK-42898][SQL] Mark that string/date casts do not need time zone id
This PR takes over https://github.com/apache/spark/pull/40524 as the original author is busy with other works.

It also close #40524

### What changes were proposed in this pull request?

This removes the need for a time zone id when casting from StringType -> DateType and DateType -> StringType.

### Why are the changes needed?

It is mostly for consistency with what the code is actually doing. Marking it as needing A time zone id has no real impact to the actual execution.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass GA. And `HivePartitionFilteringSuite` is updated to adapt to the change.

Closes #42089 from pan3793/SPARK-42898.

Lead-authored-by: Robert (Bobby) Evans <bobby@apache.org>
Co-authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>