[SPARK-44505][SQL] Provide override for columnar support in Scan for DSv2
### What changes were proposed in this pull request?
Previously, when a new DSv2 data source is implemented during planning, it will always call `BatchScanExec:supportsColumnar` which will in turn iterate over all input partitions to check if they support columnar or not.

When the `planInputPartitions` method is expensive this can be problematic. This patch adds an option to the Scan interface that allows specifying a default value. For backward compatibility the default value provided by the Scan interface is partition defined, but a Scan can change it accordingly.

To fully support the changes of this PR, the following additional changes had to be done:

* `DataSourceV2ScanExecBase::outputPartitioning` removed the case for single partitions.
* `lazyval DataSourceV2ScanExecBase::groupedPartitions` added a special check for empty key group partitioning so that the simple case does not trigger a materialization of the input partitions during planning.

Additionally:
* Fixes similar issues as https://github.com/apache/spark/pull/40004

### Why are the changes needed?
Avoid costly operations during explain operations.

### Does this PR introduce _any_ user-facing change?
Np

### How was this patch tested?
Added new UT.

Closes #42099 from grundprinzip/SPARK-44505.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>