[SPARK-44425][CONNECT] Validate that user provided sessionId is an UUID
### What changes were proposed in this pull request?

We want to validate that user provided sessionId is an UUID. Existing Spark Connect python and scala clients already do that, we would like to depend on it being in this format moving forward, just like we already validate that operatoinId is an UUID.

### Why are the changes needed?

Validate what's already assumed.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing CI.

Closes #42150 from juliuszsompolski/SPARK-44425.

Authored-by: Juliusz Sompolski <julek@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>