[SPARK-44538][CONNECT][SQL] Reinstate Row.jsonValue and friends
### What changes were proposed in this pull request?
This PR reinstates `Row.jsonValue`, `Row.prettyJson` & `Row.json`.

### Why are the changes needed?
`Row.jsonValue` and other json rendering is used in a couple of hard to remove places and there are also a couple of users using it. In order to support this we needed to move quite a few date time related classes.  We were going to move them anyway because they are also needed to decouple direct arrow encoding from catalyst.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

Closes #42164 from hvanhovell/SPARK-44538.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>