[SPARK-44432][SS][CONNECT] Terminate streaming queries when a session times out in Spark Connect
### What changes were proposed in this pull request?

Spark Connect server keeps a session id at connect server alive as long as a streaming query is alive. This may not be desirable. If the client disconnects long enough and times out in server, we should let the session mapping to be removed and terminate the streaming queries.

### Why are the changes needed?

As a Spark Connect semantic, we want to cleanup running queries when client is disconnected/timed out for a long time that the related session mapping is removed from the server.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Updated existing unit test to include new test case.

Closes #42193 from bogao007/query-termination.

Authored-by: bogao007 <bo.gao@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>