[SPARK-44584][CONNECT] Set client_type information for AddArtifactsRequest and ArtifactStatusesRequest in Scala Client
### What changes were proposed in this pull request?

Added `client_type` information to the `AddArtifactsRequest` and `ArtifactStatusesRequest` by having `ArtifactManager` take in the `SparkConnectClient.Configuration` param instead of just `userContext` (the configuration provides both the `userContext` and the `userAgent` (i.e the client type))

### Why are the changes needed?

The client_type information is missing for the `AddArtifactsRequest` and `ArtifactStatusesRequest` compared to the other requests (e.g `analyze`, `execute` etc).

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

N/A

Closes #42209 from vicennial/SPARK-44584.

Authored-by: vicennial <venkata.gudesa@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>