[SPARK-44610][SQL] DeduplicateRelations should retain Alias metadata when creating a new instance
### What changes were proposed in this pull request?

This fixes a small but obvious mistake in the rule `DeduplicateRelations`. When creating a new instance of an existing Alias, we should only make it have a new exprId, and retain everything else, including metadata.

### Why are the changes needed?

To fix a potential bug. I can't found any end-to-end query that is affected by this bug, but it's wrong to lose the metadata.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

existing tests

Closes #42242 from cloud-fan/mm.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>