[SPARK-44480][SS] Use thread pool to perform maintenance activity for hdfs/rocksdb state store providers
### What changes were proposed in this pull request?

Maintenance tasks on StateStore was being done by a single background thread, which is prone to straggling. In this change, the single background thread would instead schedule maintenance tasks to a thread pool.

By default, the maximum number of threads in the new thread pool is determined via the number of cores * 0.25, so that this thread pool doesn't take too many resources away from the query and affect performance. Users can set the number of threads explicitly via `spark.sql.streaming.stateStore.numStateStoreMaintenanceThreads`.

### Why are the changes needed?

Using a thread pool instead of a single thread for snapshotting and cleanup reduces the effect of stragglers in the background task.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Manual testing, wrote a stateful query, tracked how many maintenance tasks were done and compared this to the baseline.  `StateStoreSuite` is enough to test functional correctness and cleanup here

Closes #42066 from ericm-db/maintenance-thread-pool-optional.

Lead-authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Co-authored-by: ericm-db <132308037+ericm-db@users.noreply.github.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>