[SPARK-44658][CORE] `ShuffleStatus.getMapStatus` should return `None` instead of `Some(null)`
### What changes were proposed in this pull request?

This PR is for `master` and `branch-3.5` and aims to fix a regression due to SPARK-43043 which landed at Apache Spark 3.4.1 and reverted via SPARK-44630. This PR makes `ShuffleStatus.getMapStatus` return `None` instead of `Some(null)`.

### Why are the changes needed?

`None` is better because `Some(null)` is unsafe because it causes NPE in some cases.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs with the newly added test case.

Closes #42323 from dongjoon-hyun/SPARK-44658.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>