[SPARK-44644][PYTHON] Improve error messages for Python UDTFs with pickling errors
### What changes were proposed in this pull request?

This PR improves the error messages when a Python UDTF failed to pickle.

### Why are the changes needed?

To make the error message more user-friendly

### Does this PR introduce _any_ user-facing change?

Yes, before this PR, when a UDTF fails to pickle, it throws this confusing exception:
```
_pickle.PicklingError: Cannot pickle files that are not opened for reading: w
```
After this PR, the error is more clear:
`[UDTF_SERIALIZATION_ERROR] Cannot serialize the UDTF 'TestUDTF': Please check the stack trace and make sure that the function is serializable.`

And for spark session access inside a UDTF:
`[UDTF_SERIALIZATION_ERROR] it appears that you are attempting to reference SparkSession inside a UDTF. SparkSession can only be used on the driver, not in code that runs on workers. Please remove the reference and try again.`

### How was this patch tested?

New UTs.

Closes #42309 from allisonwang-db/spark-44644-pickling.

Authored-by: allisonwang-db <allison.wang@databricks.com>
Signed-off-by: Takuya UESHIN <ueshin@databricks.com>