[SPARK-44682][PS] Make pandas error class message_parameters strings
### What changes were proposed in this pull request?
This PR converts the types for message_parameters for pandas error classes to string, to ensure ability to compare error class messages in tests.

### Why are the changes needed?
The change ensures the ability to compare error class messages in tests.

### Does this PR introduce _any_ user-facing change?
No, the PR does not affect the user-facing view of the error messages.

### How was this patch tested?
Updated `python/pyspark/pandas/tests/test_utils.py` and existing tests

Closes #42348 from asl3/string-pandas-error-types.

Authored-by: Amanda Liu <amanda.liu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>