[SPARK-44132][SQL] Materialize `Stream` of join column names to avoid codegen failure
### What changes were proposed in this pull request?

Materialize passed join columns as an `IndexedSeq` before passing it to the lower layers.

### Why are the changes needed?

When nesting multiple full outer joins using column names which are a `Stream`, the code generator will generate faulty code resulting in a NPE or bad `UnsafeRow` access at runtime. See the 2 added test cases.

### Why are the changes needed?
Otherwise the code will crash, see the 2 added test cases.  Which show an NPE and a bad `UnsafeRow` access.

### Does this PR introduce _any_ user-facing change?
No, only bug fix.

### How was this patch tested?
A reproduction scenario was created and added to the code base.

Closes #41712 from steven-aerts/SPARK-44132-fix.

Authored-by: Steven Aerts <steven.aerts@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>